Base/large channel controllers refactored
