+++ frontend/src/app/components/channel/control/control.html
@@ -7,6 +7,7 @@
     user="vm.previewUser"
     is-ready="vm.isReadyToSwitch()"
     trigger="vm.trigger()"
+    hotkey="vm.hotkey"
     show-audio-toggle></slot-widget>
 
   <md-autocomplete
